Initialize CodeBreaker
Initialize is empty since there are no instance variables that I can
think of for now. make_guess method randomizes a guess for now.

GUESS_COLORS and HINT_COLORS are put into Game since they are not
aspects of the board itself and do not belong exclusively to either
player class.