Add input validation for make_guess
Created and used an array of all possible combinations. This will be
needed later in the algorithm for the CodeBreaker AI, so I decided to
make it now and use it for input validation. It may not be the most
efficient but it works quickly enough and it is easy to understand.