Write main logic for give_hint in CodeMaker
The only thing left is to get the hint on the board.

I moved the constants across to Board due to a bug with accessing a
constant from a different class. I wonder if this is because of order of
initialization, but I am unsure. Anyway, keeping them all in Board makes
sense as they can be considered part of the Board.

I also made an add_guess method in Board due to a bug where I tried to
access the board guesses which stored the colored circles, not the
character array! I use the add_guess method to update the new
latest_guess instance variable as well as add the colored circles to the
guesses to display.

The new instance variables latest_guess and latest_hint only store the
most recent guess and hint respectively since only these values are
planned to be required but this can always be updated to an array of all
values with ease.

The code_tally instance variable in CodeMaker exists so that it only has
to be calculated once and then be used many times in the sum_min_tallies
method.