Create HintCreatable module
To make the code more DRY.

This had the side effect of changing the add_hint method of Board to
deal with the hint hash as an argument instead of the red and white pins
individually. This is because the create_hint method in the new module
has to return all this information in one object so I made it directly
into the hint hash and returned it.

The sum_min_tallies method now finds the unique array of colors from
both tally hashes instead of using the GUESS_COLORS constant from
PinCreatable, in order to avoid requiring it in the new module.

I think the reduce_combinations method of ComputerCodeBreaker was not
private before, so now it is since it is not required in the interface.